Ownership check added.

+++ src/main/java/com/github/keler1024/expensesandincomeservice/service/ChangeService.java
@@ -5,6 +5,7 @@ import com.github.keler1024.expensesandincomeservice.data.entity.Change;
 import com.github.keler1024.expensesandincomeservice.data.entity.Category;
 import com.github.keler1024.expensesandincomeservice.data.entity.Tag;
 import com.github.keler1024.expensesandincomeservice.exception.ResourceNotFoundException;
+import com.github.keler1024.expensesandincomeservice.exception.UnauthorizedAccessException;
 import com.github.keler1024.expensesandincomeservice.model.converter.ChangeConverter;
 import com.github.keler1024.expensesandincomeservice.model.request.ChangeRequest;
 import com.github.keler1024.expensesandincomeservice.model.response.ChangeResponse;
@@ -89,11 +90,18 @@ public class ChangeService extends BaseService<ChangeRequest, Change, ChangeResp
         return converter.createResponses(entityRepository.findAll(specificationBuilder.build()));
     }
 
-    public ChangeResponse add(ChangeRequest changeRequest) {
-        if (changeRequest == null) {
+    @Override
+    public ChangeResponse add(ChangeRequest changeRequest, Long ownerId) {
+        if (ownerId == null || ownerId < 0 || changeRequest == null) {
             throw new IllegalArgumentException();
         }
+
         Change change = converter.convertToEntity(changeRequest);
+
+        if (!ownerId.equals(change.getAccount().getOwnerId())) {
+            throw new UnauthorizedAccessException();
+        }
+
         change.setAccount(accountRepository.findById(changeRequest.getAccountId())
                 .orElseThrow(
                         () -> new ResourceNotFoundException(
@@ -112,8 +120,8 @@ public class ChangeService extends BaseService<ChangeRequest, Change, ChangeResp
     }
 
     @Override
-    public ChangeResponse update(ChangeRequest changeRequest, Long id) {
-        if (id == null || id < 0 || changeRequest == null) {
+    public ChangeResponse update(ChangeRequest changeRequest, Long id, Long ownerId) {
+        if (id == null || id < 0 || ownerId == null || ownerId < 0 || changeRequest == null) {
             throw new IllegalArgumentException();
         }
         Change change = entityRepository.findById(id).orElseThrow(
@@ -121,6 +129,9 @@ public class ChangeService extends BaseService<ChangeRequest, Change, ChangeResp
                         String.format("Account change with id %d not found", id)
                 )
         );
+        if (!ownerId.equals(change.getAccount().getOwnerId())) {
+            throw new UnauthorizedAccessException();
+        }
         Long oldAmount = change.getAmount();
         change.setAmount(changeRequest.getAmount());
         change.setDateTime(changeRequest.getDateTime());
@@ -140,7 +151,7 @@ public class ChangeService extends BaseService<ChangeRequest, Change, ChangeResp
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long ownerId) {
         if (id == null || id < 0) {
             throw new IllegalArgumentException("Null instead of Account change id provided");
         }
@@ -150,8 +161,16 @@ public class ChangeService extends BaseService<ChangeRequest, Change, ChangeResp
                 )
         );
         Account account = change.getAccount();
+        if (!ownerId.equals(account.getOwnerId())) {
+            throw new UnauthorizedAccessException();
+        }
         account.setMoney(account.getMoney() - change.getAmount());
         accountRepository.save(account);
         entityRepository.deleteById(id);
+    }
+
+    @Override
+    protected Long getEntityOwnerId(Change entity) {
+        return entity.getAccount().getOwnerId();
     }
 }
